Atualizando a branch Iago-Lobo

+++ include/Vec2.h
@@ -7,7 +7,7 @@
 class Vec2{
 	public:
 		Vec2();
-		Vec2(int X, int Y);
+		Vec2(float X, float Y);
 		~Vec2();
 		float x;
 		float y;
